+++ Runtime/Timer.cs
@@ -19,7 +19,14 @@ namespace Nonatomic.TimerKit
 		/// <summary>
 		/// Gets the running state of the timer.
 		/// </summary>
-		public bool IsRunning => _timer.IsRunning;
+		public bool IsRunning
+		{
+			get
+			{
+				InitializeTimerIfNeeded();
+				return _timer.IsRunning;
+			}
+		}
 		
 		/// <summary>
 		/// Gets or sets the total duration of the timer in seconds.
@@ -27,29 +34,65 @@ namespace Nonatomic.TimerKit
 		/// </summary>
 		public float Duration
 		{
-			get => _timer.Duration;
-			set => _timer.Duration = value;
+			get
+			{
+				InitializeTimerIfNeeded();
+				return _timer.Duration;
+			}
+			set
+			{
+				InitializeTimerIfNeeded();
+				_timer.Duration = value;
+			}
 		}
 
 		/// <summary>
 		/// Gets the remaining time in seconds.
 		/// </summary>
-		public float TimeRemaining => _timer.TimeRemaining;
+		public float TimeRemaining
+		{
+			get
+			{
+				InitializeTimerIfNeeded();
+				return _timer.TimeRemaining;
+			}
+		}
 
 		/// <summary>
 		/// Gets the elapsed time in seconds since the timer was started.
 		/// </summary>
-		public float TimeElapsed => _timer.TimeElapsed;
+		public float TimeElapsed
+		{
+			get
+			{
+				InitializeTimerIfNeeded();
+				return _timer.TimeElapsed;
+			}
+		}
 
 		/// <summary>
 		/// Gets the progress of the timer as a fraction of the elapsed time over the total duration.
 		/// </summary>
-		public float ProgressElapsed => _timer.ProgressElapsed;
+		public float ProgressElapsed
+		{
+			get
+			{
+				InitializeTimerIfNeeded();
+				return _timer.ProgressElapsed;
+			}
+		}
 
 		/// <summary>
 		/// Gets the remaining progress of the timer as a fraction.
 		/// </summary>
-		public float ProgressRemaining => _timer.ProgressRemaining;
+		public float ProgressRemaining
+		{
+			get
+			{
+				InitializeTimerIfNeeded();
+				return _timer.ProgressRemaining;
+			}
+		}
 
 		[SerializeField] private float _duration = 10f;
 		[SerializeField] private bool _useScaledTime = true;
@@ -61,45 +104,77 @@ namespace Nonatomic.TimerKit
 		/// <summary>
 		/// Gets the time as either TimeRemaining, TimeElapsed, ProgressElapsed, ProgressRemaining
 		/// </summary>
-		public virtual float TimeByType(TimeType type) => _timer.TimeByType(type);
+		public virtual float TimeByType(TimeType type)
+		{
+			InitializeTimerIfNeeded();
+			return _timer.TimeByType(type);
+		}
 
 		/// <summary>
 		/// Starts or restarts the timer.
 		/// </summary>
-		public virtual void StartTimer() => _timer.StartTimer();
-		
+		public virtual void StartTimer()
+		{
+			InitializeTimerIfNeeded();
+			_timer.StartTimer();
+		}
+
 		/// <summary>
 		/// Resumes the timer without resetting.
 		/// </summary>
-		public virtual void ResumeTimer() => _timer.ResumeTimer();
+		public virtual void ResumeTimer()
+		{
+			InitializeTimerIfNeeded();
+			_timer.ResumeTimer();
+		}
 
 		/// <summary>
 		/// Stops the timer, pausing the countdown.
 		/// </summary>
-		public virtual void StopTimer() => _timer.StopTimer();
+		public virtual void StopTimer()
+		{
+			InitializeTimerIfNeeded();
+			_timer.StopTimer();
+		}
 
 		/// <summary>
 		/// Resets the timer to its initial state with the full duration remaining.
 		/// </summary>
-		public virtual void ResetTimer() => _timer.ResetTimer();
+		public virtual void ResetTimer()
+		{
+			InitializeTimerIfNeeded();
+			_timer.ResetTimer();
+		}
 
 		/// <summary>
 		/// Advances the timer forward by a specified amount of time.
 		/// </summary>
 		/// <param name="time">The duration in seconds to fast forward the timer.</param>
-		public virtual void FastForward(float time) => _timer.FastForward(time);
+		public virtual void FastForward(float time)
+		{
+			InitializeTimerIfNeeded();
+			_timer.FastForward(time);
+		}
 
 		/// <summary>
 		/// Rewinds the timer backward by a specified number of seconds.
 		/// </summary>
 		/// <param name="time">The amount of time, in seconds, to rewind the timer.</param>
-		public virtual void Rewind(float time) => _timer.Rewind(time);
+		public virtual void Rewind(float time)
+		{
+			InitializeTimerIfNeeded();
+			_timer.Rewind(time);
+		}
 
 		/// <summary>
 		/// Adds a milestone to the timer that will trigger a callback when a specific timer condition is met.
 		/// </summary>
-		public virtual void AddMilestone(TimerMilestone milestone) => _timer.AddMilestone(milestone);
-		
+		public virtual void AddMilestone(TimerMilestone milestone)
+		{
+			InitializeTimerIfNeeded();
+			_timer.AddMilestone(milestone);
+		}
+
 		/// <summary>
 		/// Adds a range milestone that triggers at regular intervals within a specified range.
 		/// </summary>
@@ -110,27 +185,53 @@ namespace Nonatomic.TimerKit
 		/// <param name="callback">The callback to execute at each interval</param>
 		/// <param name="isRecurring">Whether this milestone should re-trigger every time the timer restarts</param>
 		/// <returns>The created TimerRangeMilestone</returns>
-		public virtual TimerRangeMilestone AddRangeMilestone(TimeType type, float rangeStart, float rangeEnd, float interval, Action callback, bool isRecurring = false) 
-			=> _timer.AddRangeMilestone(type, rangeStart, rangeEnd, interval, callback, isRecurring);
-		
+		public virtual TimerRangeMilestone AddRangeMilestone(TimeType type, float rangeStart, float rangeEnd, float interval, Action callback, bool isRecurring = false)
+		{
+			InitializeTimerIfNeeded();
+			return _timer.AddRangeMilestone(type, rangeStart, rangeEnd, interval, callback, isRecurring);
+		}
+
 		/// <summary>
 		/// Removes a specific milestone from the timer.
 		/// </summary>
-		public virtual void RemoveMilestone(TimerMilestone milestone) => _timer.RemoveMilestone(milestone);
+		public virtual void RemoveMilestone(TimerMilestone milestone)
+		{
+			InitializeTimerIfNeeded();
+			_timer.RemoveMilestone(milestone);
+		}
 
 		/// <summary>
 		/// Removes all milestones associated with the timer. This operation clears any timing checkpoints or markers that have been added.
 		/// </summary>
-		public void RemoveAllMilestones() => _timer.RemoveAllMilestones();
-		
+		public void RemoveAllMilestones()
+		{
+			InitializeTimerIfNeeded();
+			_timer.RemoveAllMilestones();
+		}
+
 		/// <summary>
 		/// Removes milestones that meet a specific condition.
 		/// </summary>
-		public virtual void RemoveMilestonesByCondition(Predicate<TimerMilestone> condition) => _timer.RemoveMilestonesByCondition(condition);
+		public virtual void RemoveMilestonesByCondition(Predicate<TimerMilestone> condition)
+		{
+			InitializeTimerIfNeeded();
+			_timer.RemoveMilestonesByCondition(condition);
+		}
 
 		protected virtual void Awake()
 		{
-			_timer = new StandardTimer(_duration, _timeSource);
+			InitializeTimerIfNeeded();
+		}
+
+		/// <summary>
+		/// Initializes the timer if it hasn't been initialized yet.
+		/// </summary>
+		private void InitializeTimerIfNeeded()
+		{
+			if (_timer == null)
+			{
+				_timer = new StandardTimer(_duration, _timeSource);
+			}
 		}
 		
 		/// <summary>
@@ -164,6 +265,7 @@ namespace Nonatomic.TimerKit
 		
 		protected virtual void OnEnable()
 		{
+			InitializeTimerIfNeeded();
 			_timer.OnStart += HandleTimerStart;
 			_timer.OnResume += HandleTimerResume;
 			_timer.OnStop += HandleTimerStopped;
@@ -174,6 +276,8 @@ namespace Nonatomic.TimerKit
 
 		protected virtual void OnDisable()
 		{
+			if (_timer == null) return;
+
 			_timer.OnStart -= HandleTimerStart;
 			_timer.OnResume -= HandleTimerResume;
 			_timer.OnStop -= HandleTimerStopped;